fix: fall back to setting the invite url in the poll embed if the DM could not be sent
Discord disallows bots from initiating conversations with users that they have
never encountered, as such we want to gracefully fall back to setting the
generated invite url in the poll message if the DM could not be sent.

This implies that whoever started the poll, or someone in their stead, is
tasked with sharing the invite url with the invitee once the poll is closed.